+++ CHANGES.md
@@ -35,7 +35,7 @@ YYYY-MM-DD
 Speechoid: 0.1.2
 
 * [T279412](https://phabricator.wikimedia.org/T279412) Add link to help page for Special:EditLexicon
-* [T339840](https://phabricator.wikimedia.org/T339840) Editor: Redirect logged out users to login page if they lack right to edit lexicon
+* [T339840](https://phabricator.wikimedia.org/T339840) Enable edit button on consumer wiki
 * [T276332](https://phabricator.wikimedia.org/T276332) Editor: disable preview to indicate that the system is busy
 * [T276332](https://phabricator.wikimedia.org/T276332) Simplify message for error when mapping transcriptions
 * [T279412](https://phabricator.wikimedia.org/T279412) Change label of submit button in lexicon editor to reflect what it does